hangman
In this project, I have made hangman game using Python. Hangman is a paper and pencil guessing game. Player thinks of a word, phrase or sentence and tries to guess it by suggesting letters or numbers, within a certain number of guesses. Player has the length of the word that have to guess.